Refactor: Split profile management utils into modules
This commit refactors the profile management utility functions into smaller, more manageable modules. This improves code organization and maintainability. Unused imports and files are also removed.
