+++ src/utils/profile/ProfileError.ts
@@ -1,0 +1,13 @@
+
+/**
+ * Custom error class for profile-related errors
+ */
+export class ProfileError extends Error {
+  constructor(
+    message: string,
+    public readonly type: 'not_found' | 'network' | 'unknown'
+  ) {
+    super(message);
+    this.name = 'ProfileError';
+  }
+}
