+++ src/utils/profile/index.ts
@@ -1,0 +1,7 @@
+
+// Re-export everything from the profile module
+export { ProfileError } from './ProfileError';
+export { fetchProfile } from './ProfileFetcher';
+export { updateProfile } from './ProfileUpdater';
+export { fetchRolePermissions, fetchPlanDetails } from './RolePermissionsService';
+export type { UserProfile, UserRole, SubscriptionPlan, UserPermission, SystemRole, SystemPlan } from './types';
