+++ src/utils/profile/validators.ts
@@ -1,0 +1,32 @@
+
+import { SubscriptionPlan } from './types';
+
+/**
+ * Helper to validate subscription plan
+ */
+export const validatePlan = (plan: string | null): SubscriptionPlan | undefined => {
+  return plan || undefined;
+};
+
+/**
+ * Helper to validate notification preferences
+ */
+export const validateNotificationPreferences = (preferences: unknown) => {
+  const defaultPreferences = {
+    travel_alerts: true,
+    policy_changes: true,
+    documentation_reminders: true
+  };
+
+  if (!preferences || typeof preferences !== 'object') {
+    return defaultPreferences;
+  }
+
+  const prefs = preferences as Record<string, unknown>;
+  
+  return {
+    travel_alerts: typeof prefs.travel_alerts === 'boolean' ? prefs.travel_alerts : defaultPreferences.travel_alerts,
+    policy_changes: typeof prefs.policy_changes === 'boolean' ? prefs.policy_changes : defaultPreferences.policy_changes,
+    documentation_reminders: typeof prefs.documentation_reminders === 'boolean' ? prefs.documentation_reminders : defaultPreferences.documentation_reminders
+  };
+};
